split out NextIndex Cache analyzer
